+++ ru.taximaxim.codekeeper.core/src/main/java/ru/taximaxim/codekeeper/core/Consts.java
@@ -46,6 +46,8 @@ public interface Consts {
 
     String PLUGIN_ID = "ru.taximaxim.codekeeper.core";
 
+    String POOL_SIZE = "ru.taximaxim.codekeeper.parser.poolsize";
+
     String FILENAME_WORKING_DIR_MARKER = ".pgcodekeeper";
     String VERSION_PROP_NAME = "version";
     String EXPORT_CURRENT_VERSION = "0.6.0";
@@ -69,7 +71,7 @@ public interface Consts {
     String DUPLICATE_RELATION = "'42P07'";
     //If object exists we get this error code.
     String DUPLICATE_OBJECT = "'42710'";
-    
+
     /**
      * Log message when in the SELECT part aliases is duplicate
      */