+++ ru.taximaxim.codekeeper.ui/src/ru/taximaxim/codekeeper/ui/UiUtils.java
@@ -15,6 +15,8 @@
  *******************************************************************************/
 package ru.taximaxim.codekeeper.ui;
 
+import java.util.Locale;
+
 import org.eclipse.jface.viewers.IStructuredSelection;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.events.KeyEvent;
@@ -28,6 +30,15 @@ public final class UiUtils {
     public static boolean isCommandCopy(KeyEvent e) {
         int key = e.stateMask;
         return (((key & SWT.CTRL) == SWT.CTRL) || ((key & SWT.COMMAND) == SWT.COMMAND)) && e.keyCode == 'c';
+    }
+
+    public static String getOS() {
+        return System.getProperty("os.name", "generic").toLowerCase(Locale.ROOT);
+    }
+
+    public static boolean isMac() {
+        String os = getOS();
+        return os.indexOf("mac") >= 0 || os.indexOf("darwin") >= 0;
     }
 
     private UiUtils() {