+++ ru.taximaxim.codekeeper.ui/src/ru/taximaxim/codekeeper/ui/prefs/GeneralPrefPage.java
@@ -24,17 +24,23 @@ import org.eclipse.swt.events.SelectionEvent;
 import org.eclipse.swt.layout.GridData;
 import org.eclipse.swt.widgets.Button;
 import org.eclipse.swt.widgets.Label;
+import org.eclipse.swt.widgets.Text;
 import org.eclipse.ui.IWorkbench;
 import org.eclipse.ui.IWorkbenchPreferencePage;
 
+import ru.taximaxim.codekeeper.core.Consts;
 import ru.taximaxim.codekeeper.core.parsers.antlr.AntlrParser;
 import ru.taximaxim.codekeeper.ui.Activator;
+import ru.taximaxim.codekeeper.ui.EclipseIniWriter;
+import ru.taximaxim.codekeeper.ui.IntegerVerifyListener;
 import ru.taximaxim.codekeeper.ui.UIConsts.PREF;
 import ru.taximaxim.codekeeper.ui.localizations.Messages;
 import ru.taximaxim.codekeeper.ui.pgdbproject.parser.PgDbParser;
 
-public class GeneralPrefPage extends FieldEditorPreferencePage
+public final class GeneralPrefPage extends FieldEditorPreferencePage
         implements IWorkbenchPreferencePage {
+
+    private Text threadsTxt;
 
     public GeneralPrefPage() {
         super(GRID);
@@ -91,6 +97,14 @@ public class GeneralPrefPage extends FieldEditorPreferencePage
         new Label(getFieldEditorParent(), SWT.SEPARATOR | SWT.HORIZONTAL)
             .setLayoutData(new GridData(SWT.FILL, SWT.DEFAULT, true, false, 2, 1));
 
+        new Label(getFieldEditorParent(), SWT.NONE)
+            .setText(Messages.GeneralPrefPage_number_of_loading_threads);
+
+        threadsTxt = new Text(getFieldEditorParent(), SWT.BORDER);
+        threadsTxt.setLayoutData(new GridData(SWT.FILL, SWT.CENTER, true, false));
+        threadsTxt.setText(Integer.toString(AntlrParser.getPoolSize()));
+        threadsTxt.addVerifyListener(new IntegerVerifyListener());
+
         addField(new BooleanFieldEditor(PREF.HEAP_SIZE_WARNING,
                 Messages.GeneralPrefPage_alert_if_heap_size_less_than_necessary,
                 getFieldEditorParent()));
@@ -110,5 +124,18 @@ public class GeneralPrefPage extends FieldEditorPreferencePage
                 PgDbParser.cleanAll();
             }
         });
+    }
+
+    @Override
+    public boolean performOk() {
+        String newThreadVal = threadsTxt.getText();
+
+        if (AntlrParser.getPoolSize() != Integer.parseInt(newThreadVal)) {
+            var fullPropertyName = "-D" + Consts.POOL_SIZE + '='; //$NON-NLS-1$
+            EclipseIniWriter.write(getShell(), fullPropertyName, newThreadVal,
+                    Messages.GeneralPrefPage_number_of_thread);
+        }
+
+        return super.performOk();
     }
 }